+++ console_backend/src/errors.rs
@@ -1,11 +1,10 @@
 pub(crate) const HEARTBEAT_LOCK_MUTEX_FAILURE: &str = "unable to lock heartbeat mutex";
-pub(crate) const SHARED_STATE_LOCK_MUTEX_FAILURE: &str = "unable to lock shared_state mutex";
+pub const SHARED_STATE_LOCK_MUTEX_FAILURE: &str = "unable to lock shared_state mutex";
 pub(crate) const UPDATE_STATUS_LOCK_MUTEX_FAILURE: &str = "unable to lock update status mutex";
 pub(crate) const CAP_N_PROTO_SERIALIZATION_FAILURE: &str = "unable to serialize capnproto message";
 #[allow(dead_code)]
-pub(crate) const CAP_N_PROTO_DESERIALIZATION_FAILURE: &str =
-    "unable to deserialize capnproto message";
-pub(crate) const CONVERT_TO_STR_FAILURE: &str = "error converting to str";
+pub const CAP_N_PROTO_DESERIALIZATION_FAILURE: &str = "unable to deserialize capnproto message";
+pub const CONVERT_TO_STR_FAILURE: &str = "error converting to str";
 pub(crate) const GET_MUT_OBJECT_FAILURE: &str = "error trying to get mut object";
 pub(crate) const UNABLE_TO_STOP_TIMER_THREAD_FAILURE: &str = "unable to kill running timer thread";
 pub(crate) const UNABLE_TO_SEND_INS_UPDATE_FAILURE: &str = "unable to send an ins status update";
